Change: use '*' in [string] when not used.
If the user does not use '*' in the [string] to
search then it would be added at the start and
end of [string]. This makes more easy to search
for an installed package rather than use it's
entire name. It also makes the script to has
the same behaviour as sbo_find.